requirements.txt
google-genai
streamlit
fastapi
python-dotenv
jinja2
uvicorn
